Restructure, integrate allora sdk and add function for retrieving price predictions
